instead of creating branch names randomly, use a hash of the PR url
the other idea I considered was adding some kind of database to store
the random branch names for each PR. but that would add more moving
parts.

I'm not really sure hashing is a good idea because the mantainer will
be able to enumerate all PRs to the anonydog fork and test the hashes.
On the other hand, are we actually trying to protect against this
kind of determined mantainer?